Seperate tox env for functional tests (#25)
Seperate tox env for unit and functional tests.
Functional tests require access to openstack cluster
and are not part of ci yet.
Update docs with info on running functional tests locally
