+++ tox.ini
@@ -6,6 +6,7 @@ envlist =
     packaging
     py{36,37,38,39}
     py{36,37,38,39}-{devel}
+    py{36,37,38,39}-functional
 
 # do not enable skip missing to avoid CI false positives
 skip_missing_interpreters = False
@@ -23,8 +24,48 @@ deps =
     py{36,37,38,39}: molecule[test]
     py{36,37,38,39}-{devel}: git+https://github.com/ansible-community/molecule.git@master#egg=molecule[test]
 commands =
-    pytest --collect-only
-    pytest --color=yes {tty:-s}
+    pytest molecule_openstack/test/unit --collect-only
+    pytest molecule_openstack/test/unit --color=yes {tty:-s}
+setenv =
+    PIP_DISABLE_PIP_VERSION_CHECK=1
+    PY_COLORS={env:PY_COLORS:1}
+    # pip: Avoid 2020-01-01 warnings: https://github.com/pypa/pip/issues/6207
+    PYTHONWARNINGS=ignore:DEPRECATION::pip._internal.cli.base_command
+    PYTHONDONTWRITEBYTECODE=1
+    # This should pass these args to molecule, no effect here as this is the default
+    # but it validates that it accepts extra params.
+    MOLECULE_OPTS=--destroy always
+passenv =
+    CI
+    CURL_CA_BUNDLE
+    PYTEST_OPTIONS
+    REQUESTS_CA_BUNDLE
+    SSH_AUTH_SOCK
+    SSL_CERT_FILE
+    TOXENV
+    TWINE_*
+whitelist_externals =
+    bash
+    twine
+    pytest
+    pre-commit
+
+[testenv:py{36,37,38,39}-functional]
+description =
+    Functional testing, require access to openstack cluster
+usedevelop = True
+# download assures tox gets newest pip, see https://github.com/tox-dev/tox/issues/791
+download = true
+# sitepackages = True
+extras = test
+deps =
+    molecule[test]
+    ansible-base >=2.10.0
+    ansible >= 2.10.0
+    openstacksdk >= 0.50.0
+commands =
+    pytest molecule_openstack/test/functional --collect-only
+    pytest molecule_openstack/test/functional --color=yes {tty:-s}
 setenv =
     ANSIBLE_FORCE_COLOR={env:ANSIBLE_FORCE_COLOR:1}
     ANSIBLE_INVENTORY={toxinidir}/tests/hosts.ini
@@ -54,6 +95,8 @@ passenv =
     TRAVIS
     TRAVIS_*
     TWINE_*
+    MOLECULE_*
+    OS_*
 whitelist_externals =
     bash
     twine
